SFORM-77 Service部分実装

+++ app/net/macolabo/sform2/services/ApiToken/insert/ApiTokenInsertRequest.scala
@@ -3,5 +3,5 @@ package net.macolabo.sform2.services.ApiToken.insert
 case class ApiTokenInsertRequest(
                                 group_name: String,
                                 token: String,
-                                expiry_days: Int
+                                expiry_days: Long
                                 )
